Improve the way we start service
We only need to set action for starting service in intent,
and once the service is launched, it will collect necessary user
preferences on its own.
This will make it easier to start service from difference parts of
the app, such as boot completed receiver and qs tile(which will be
added later).